backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-filter
django-allauth
mysqlclient
razorpay
python-dotenv
Pillow
